+++ lib/private/StormByte/logger/implementation.cxx
@@ -714,4 +714,5 @@ namespace StormByte::Logger {
 	template STORMBYTE_LOGGER_PUBLIC Implementation& Implementation::operator<<<const wchar_t*>(const wchar_t* const& value);
 	template STORMBYTE_LOGGER_PUBLIC Implementation& Implementation::operator<<<std::string_view>(const std::string_view& value);
 	template STORMBYTE_LOGGER_PUBLIC Implementation& Implementation::operator<<<std::wstring_view>(const std::wstring_view& value);
+	template STORMBYTE_LOGGER_PUBLIC Implementation& Implementation::operator<<<std::span<const std::byte>>(const std::span<const std::byte>& value);
 }